Update 14 -master-password made

+++ SecurePassword/SecurePassword/Backend(Art)/PasswordGenerator/IPasswordGenerator.cs
@@ -8,7 +8,7 @@ namespace SecurePassword;
 
 internal interface IPasswordGenerator
 {
-        abstract public static string GeneratePassword(bool useLowercase, bool useUppercase, bool useDigits, bool useSpecial, short passwordLength); // генерация пароля с указанием длины
+        abstract public static string GeneratePassword(bool useLowercase, bool useUppercase, bool useDigits, bool useSpecial, byte passwordLength); // генерация пароля с указанием длины
         abstract public static string GeneratePassword(bool useLowercase, bool useUppercase, bool useDigits, bool useSpecial); // генерация пароля без указания длины
         abstract public static bool ValidatePassword(string password,bool useLowercase,bool useUppercase,bool useDigits,bool useSpecial); // валидация пароля
